update anopheles to 7165 instead of 180454
Former-commit-id: 3464c46012bb17a016b105223987936c4ac0a239

+++ bio/sources/pubmed-gene/main/src/org/intermine/bio/dataconversion/PubMedGeneConverter.java
@@ -84,7 +84,7 @@ public class PubMedGeneConverter extends FileConverter
         // D.melanogaster
         organismsToProcess.add(new Integer(7227));
         // A.gambiae str PEST
-        organismsToProcess.add(new Integer(180454));
+        organismsToProcess.add(new Integer(9165));
 
         Item datasource = createItem("DataSource");
         datasource.setAttribute("name", "NCBI");
